requirements.txt
streamlit==1.31.1
python-dotenv==0.21.0
google-generativeai==0.4.0
pillow==10.3.0